requirements.txt
aiogram==3.18.0
openai==1.75.0
asyncpg==0.30.0
python-dotenv==1.1.0